chore(fluxWizard): refactor API files
